change var i from unsigned int to long unsigned int

+++ 0x13-bit_manipulation/1-print_binary.c
@@ -6,7 +6,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned int i = n;
+	unsigned long int i = n;
 	int cnt = 0;
 
 	if (n == 0)
